Small updates to WQ prep
It threw me off to have the TargetLatitude and Latitude referenced in the same code chunk (Loading WQ data). 
Moved Select() to be the first operator, so that all following would be acting on the same truncated and renamed dataframe. 

Enacted these changes directly through GitHub rather than cloning the repository to my computer. Tested in a temp markdown on my desktop and it worked well.