Revised code for second attempt
Output of this branch resulted in a smaller dataset without expanded date frame.  Reviewed code for errors - found one. Revised code to look for sources, but think it could have been an issue with the datasets used as input. Will ask someone with access to tox box to attempt re-run.